changed session, allowed student to decline admission

+++ routes/web.php
@@ -242,9 +242,9 @@ Route::prefix('student')
     ->middleware('theme:dashboard')
     ->name('student.')->group(function () {
 
-        Route::get('/select-session/{user_id}', [StudentOperation::class, 'select_session_view']);
-        Route::post('/select-session/{user_id}', [StudentOperation::class, 'confirm_session'])->name('select-session');
-        Route::delete('/delete-student-admission/{user_id}', [StudentOperation::class, 'delete_admission'])->name('delete-student-admission');
+        Route::get('/select-session/{user_id}', [StudentOperation::class, 'select_session_view'])->middleware('auth');
+        Route::post('/select-session/{user_id}', [StudentOperation::class, 'confirm_session'])->name('select-session')->middleware('auth');
+        Route::delete('/delete-student-admission/{user_id}', [StudentOperation::class, 'delete_admission'])->name('delete-student-admission')->middleware('auth');
 
         Route::middleware(['auth:web'])->group(function () {
             Route::get('/dashboard', [StudentOperation::class, 'dashboard'])->name('dashboard');
